requirements.txt
gradio
pandas
openai
loguru
promptic
tenacity
bs4
pydub
fastapi
uvicorn
python-dotenv